Refactor MainSettings to a folder.
