--- app/src/main/java/com/bestCatHustlers/sukodublitz/MainSettingsContract.java
@@ -1,24 +1,0 @@
-package com.bestCatHustlers.sukodublitz;
-
-import android.content.SharedPreferences;
-
-public interface MainSettingsContract {
-    interface View {
-
-    }
-    interface Presenter{
-        void turnSound(boolean on);
-        boolean getSound();
-        void turnMusic(boolean on);
-        boolean getMusic();
-        void viewDestroy();
-    }
-    interface Model {
-        void setSound(boolean on);
-        void setMusic(boolean on);
-        boolean getSound();
-        boolean getMusic();
-
-        void restoreState(SharedPreferences settings);
-    }
-}
